Implement background tasks for data ingestion and fix /register API

+++ loans/models.py
@@ -1,26 +1,26 @@
 from django.db import models
 
 class Customer(models.Model):
-    customer_id = models.IntegerField(primary_key=True) 
+    customer_id = models.AutoField(primary_key=True)
     first_name = models.CharField(max_length=100)
     last_name = models.CharField(max_length=100)
     age = models.IntegerField() 
     phone_number = models.CharField(max_length=15)
-    monthly_salary = models.DecimalField(max_digits=10, decimal_places=2)
-    approved_limit = models.DecimalField(max_digits=12, decimal_places=2)
-    current_debt = models.DecimalField(max_digits=12, decimal_places=2)
+    monthly_salary = models.IntegerField()
+    approved_limit = models.IntegerField()
+    current_debt = models.IntegerField()
 
     def __str__(self):
         return f"{self.first_name} {self.last_name}"
 
 
 class Loan(models.Model):
-    loan_id = models.IntegerField(primary_key=True)
+    loan_id = models.AutoField(primary_key=True)
     customer = models.ForeignKey(Customer, on_delete=models.CASCADE) 
-    loan_amount = models.DecimalField(max_digits=12, decimal_places=2)
+    loan_amount = models.FloatField()
     tenure = models.IntegerField()
-    interest_rate = models.DecimalField(max_digits=5, decimal_places=2)
-    monthly_repayment = models.DecimalField(max_digits=10, decimal_places=2)
+    interest_rate = models.FloatField()
+    monthly_repayment = models.FloatField()
     emis_paid_on_time = models.IntegerField()
     start_date = models.DateField()
     end_date = models.DateField()
